break in to pieces
